クラスをまとめる
