+++ ami/tensorboard_logger/tensorboard_logger.py
@@ -1,3 +1,5 @@
+import math
+import time
 from abc import ABC, abstractmethod
 from collections.abc import Mapping, MutableSequence
 from typing import Any
@@ -62,3 +64,51 @@ class TensorBoardLogger(ABC):
             self._expand_dict(hparams),
             {"hp_params": -1} if metrics is None else self._expand_dict(metrics),
         )
+
+
+class TimeIntervalLogger(TensorBoardLogger):
+    def __init__(
+        self,
+        log_dir: str,
+        log_every_n_seconds: float = 0.0,
+        **tensorboard_kwds: Any,
+    ):
+        super().__init__(log_dir, **tensorboard_kwds)
+        self.log_every_n_seconds = log_every_n_seconds
+        self.global_step = 0
+        self.previous_log_time: float = -math.inf
+
+    @property
+    def log_available(self) -> bool:
+        return time.perf_counter() - self.previous_log_time > self.log_every_n_seconds
+
+    def update(self) -> None:
+        self.global_step += 1
+        self.previous_log_time = time.perf_counter()
+
+    def log(self, tag: str, scalar: Tensor | float | int) -> None:
+        if self.log_available:
+            self.tensorboard.add_scalar(tag, scalar, self.global_step)
+
+
+class StepIntervalLogger(TensorBoardLogger):
+    def __init__(
+        self,
+        log_dir: str,
+        log_every_n_steps: int = 1,
+        **tensorboard_kwds: Any,
+    ):
+        super().__init__(log_dir, **tensorboard_kwds)
+        self.log_every_n_steps = log_every_n_steps
+        self.global_step = 0
+
+    @property
+    def log_available(self) -> bool:
+        return self.global_step % self.log_every_n_steps == 0
+
+    def update(self) -> None:
+        self.global_step += 1
+
+    def log(self, tag: str, scalar: Tensor | float | int) -> None:
+        if self.log_available:
+            self.tensorboard.add_scalar(tag, scalar, self.global_step)
